pyproject.toml
[tool.poetry]
name = "drf-template"
version = "0.1.0"
description = "Python service"
authors = ["Yakser <iaksanov2044@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.7"
python-decouple = "^3.8"
django-cors-headers = "^4.3.1"
pre-commit = "^3.6.0"
djangorestframework-simplejwt = "^5.3.1"
django-stubs = {extras = ["compatible-mypy"], version = "^4.2.7"}
djangorestframework-stubs = {extras = ["compatible-mypy"], version = "^3.14.5"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
